requirements.txt
mysql-connector-python>=8.0.26
Pillow>=9.0.0
tkinter
